turla/Resources/Penquin/requirements.txt
pycryptodome==3.15.0
pycryptodomex==3.11.0
scapy==2.4.4